# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Team Task Manager API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export type UserRole = (typeof UserRole)[keyof typeof UserRole];

export const UserRole = {
  admin: "admin",
  member: "member",
} as const;

export interface User {
  id: number;
  clerkId: string;
  email: string;
  name: string;
  /** @nullable */
  avatarUrl?: string | null;
  role: UserRole;
  createdAt: string;
}

export interface UpdateUserBody {
  name?: string;
  /** @nullable */
  avatarUrl?: string | null;
}

export type UpdateUserRoleBodyRole =
  (typeof UpdateUserRoleBodyRole)[keyof typeof UpdateUserRoleBodyRole];

export const UpdateUserRoleBodyRole = {
  admin: "admin",
  member: "member",
} as const;

export interface UpdateUserRoleBody {
  role: UpdateUserRoleBodyRole;
}

export type AdminCreateUserBodyRole =
  (typeof AdminCreateUserBodyRole)[keyof typeof AdminCreateUserBodyRole];

export const AdminCreateUserBodyRole = {
  admin: "admin",
  member: "member",
} as const;

export interface AdminCreateUserBody {
  email: string;
  firstName: string;
  lastName: string;
  password: string;
  role: AdminCreateUserBodyRole;
}

export type ProjectStatus = (typeof ProjectStatus)[keyof typeof ProjectStatus];

export const ProjectStatus = {
  active: "active",
  archived: "archived",
  completed: "completed",
} as const;

export interface Project {
  id: number;
  name: string;
  /** @nullable */
  description?: string | null;
  status: ProjectStatus;
  ownerId: number;
  color: string;
  taskCount: number;
  memberCount: number;
  completedTaskCount: number;
  createdAt: string;
  updatedAt: string;
}

export interface CreateProjectBody {
  name: string;
  /** @nullable */
  description?: string | null;
  color: string;
}

export type UpdateProjectBodyStatus =
  (typeof UpdateProjectBodyStatus)[keyof typeof UpdateProjectBodyStatus];

export const UpdateProjectBodyStatus = {
  active: "active",
  archived: "archived",
  completed: "completed",
} as const;

export interface UpdateProjectBody {
  name?: string;
  /** @nullable */
  description?: string | null;
  status?: UpdateProjectBodyStatus;
  color?: string;
}

export type ProjectMemberRole =
  (typeof ProjectMemberRole)[keyof typeof ProjectMemberRole];

export const ProjectMemberRole = {
  admin: "admin",
  member: "member",
} as const;

export interface ProjectMember {
  id: number;
  projectId: number;
  userId: number;
  role: ProjectMemberRole;
  user: User;
  joinedAt: string;
}

export type AddProjectMemberBodyRole =
  (typeof AddProjectMemberBodyRole)[keyof typeof AddProjectMemberBodyRole];

export const AddProjectMemberBodyRole = {
  admin: "admin",
  member: "member",
} as const;

export interface AddProjectMemberBody {
  userId: number;
  role: AddProjectMemberBodyRole;
}

export type UpdateMemberRoleBodyRole =
  (typeof UpdateMemberRoleBodyRole)[keyof typeof UpdateMemberRoleBodyRole];

export const UpdateMemberRoleBodyRole = {
  admin: "admin",
  member: "member",
} as const;

export interface UpdateMemberRoleBody {
  role: UpdateMemberRoleBodyRole;
}

export type TaskStatus = (typeof TaskStatus)[keyof typeof TaskStatus];

export const TaskStatus = {
  todo: "todo",
  in_progress: "in_progress",
  in_review: "in_review",
  done: "done",
} as const;

export type TaskPriority = (typeof TaskPriority)[keyof typeof TaskPriority];

export const TaskPriority = {
  low: "low",
  medium: "medium",
  high: "high",
  urgent: "urgent",
} as const;

export interface ProjectSummary {
  id: number;
  name: string;
  color: string;
}

export interface Task {
  id: number;
  title: string;
  /** @nullable */
  description?: string | null;
  status: TaskStatus;
  priority: TaskPriority;
  projectId: number;
  /** @nullable */
  assigneeId?: number | null;
  creatorId: number;
  /** @nullable */
  dueDate?: string | null;
  /** @nullable */
  completedAt?: string | null;
  project: ProjectSummary;
  assignee?: User | null;
  creator: User;
  createdAt: string;
  updatedAt: string;
}

export type CreateTaskBodyStatus =
  (typeof CreateTaskBodyStatus)[keyof typeof CreateTaskBodyStatus];

export const CreateTaskBodyStatus = {
  todo: "todo",
  in_progress: "in_progress",
  in_review: "in_review",
  done: "done",
} as const;

export type CreateTaskBodyPriority =
  (typeof CreateTaskBodyPriority)[keyof typeof CreateTaskBodyPriority];

export const CreateTaskBodyPriority = {
  low: "low",
  medium: "medium",
  high: "high",
  urgent: "urgent",
} as const;

export interface CreateTaskBody {
  title: string;
  /** @nullable */
  description?: string | null;
  status: CreateTaskBodyStatus;
  priority: CreateTaskBodyPriority;
  projectId: number;
  /** @nullable */
  assigneeId?: number | null;
  /** @nullable */
  dueDate?: string | null;
}

export type UpdateTaskBodyStatus =
  (typeof UpdateTaskBodyStatus)[keyof typeof UpdateTaskBodyStatus];

export const UpdateTaskBodyStatus = {
  todo: "todo",
  in_progress: "in_progress",
  in_review: "in_review",
  done: "done",
} as const;

export type UpdateTaskBodyPriority =
  (typeof UpdateTaskBodyPriority)[keyof typeof UpdateTaskBodyPriority];

export const UpdateTaskBodyPriority = {
  low: "low",
  medium: "medium",
  high: "high",
  urgent: "urgent",
} as const;

export interface UpdateTaskBody {
  title?: string;
  /** @nullable */
  description?: string | null;
  status?: UpdateTaskBodyStatus;
  priority?: UpdateTaskBodyPriority;
  /** @nullable */
  assigneeId?: number | null;
  /** @nullable */
  dueDate?: string | null;
}

export type DashboardSummaryTasksByStatus = {
  todo: number;
  in_progress: number;
  in_review: number;
  done: number;
};

export type DashboardSummaryTasksByPriority = {
  low: number;
  medium: number;
  high: number;
  urgent: number;
};

export interface DashboardSummary {
  totalProjects: number;
  activeProjects: number;
  totalTasks: number;
  myTasks: number;
  completedTasks: number;
  overdueTasks: number;
  tasksByStatus: DashboardSummaryTasksByStatus;
  tasksByPriority: DashboardSummaryTasksByPriority;
}

export type ActivityItemType =
  (typeof ActivityItemType)[keyof typeof ActivityItemType];

export const ActivityItemType = {
  task_created: "task_created",
  task_updated: "task_updated",
  task_completed: "task_completed",
  project_created: "project_created",
  member_added: "member_added",
} as const;

export interface ActivityItem {
  id: number;
  type: ActivityItemType;
  description: string;
  userId: number;
  userName: string;
  /** @nullable */
  projectId?: number | null;
  /** @nullable */
  projectName?: string | null;
  /** @nullable */
  taskId?: number | null;
  /** @nullable */
  taskTitle?: string | null;
  createdAt: string;
}

export type ListTasksParams = {
  projectId?: number;
  assigneeId?: number;
  status?: ListTasksStatus;
  priority?: ListTasksPriority;
};

export type ListTasksStatus =
  (typeof ListTasksStatus)[keyof typeof ListTasksStatus];

export const ListTasksStatus = {
  todo: "todo",
  in_progress: "in_progress",
  in_review: "in_review",
  done: "done",
} as const;

export type ListTasksPriority =
  (typeof ListTasksPriority)[keyof typeof ListTasksPriority];

export const ListTasksPriority = {
  low: "low",
  medium: "medium",
  high: "high",
  urgent: "urgent",
} as const;